chore: divide DB functions in new files
